refactor launch analysis button
